Move the group player into a separate folder so we can re-use it
