+++ src/components/GroupPlayer/index.ts
@@ -1,0 +1,1 @@
+export * from "./GroupPlayer";
